@@ -28,43 +28,43 @@ namespace Repositories.Repositories;
 
 internal abstract class RepositoryDecorator<T> : IRepository<T> where T : IExprEntity
 {
-	private readonly IRepository<T> _repository;
+    private readonly IRepository<T> _repository;
 
-	protected RepositoryDecorator(IRepository<T> repository) => _repository = repository;
+    protected RepositoryDecorator(IRepository<T> repository) => _repository = repository;
 
-	public virtual Task<T?> FindAsync(Expression<Func<T, bool>>? where = null,
-		IEnumerable<Expression<Func<T, object>>>? orderBy = null, int? offset = null, int? count = null,
-		CancellationToken cancellationToken = default) =>
-		_repository.FindAsync(where, orderBy, offset, count, cancellationToken);
+    public virtual Task<T?> FindAsync(Expression<Func<T, bool>>? where = null,
+        IEnumerable<Expression<Func<T, object>>>? orderBy = null, int? offset = null, int? count = null,
+        CancellationToken cancellationToken = default) =>
+        _repository.FindAsync(where, orderBy, offset, count, cancellationToken);
 
-	public virtual Task<T> GetAsync(Expression<Func<T, bool>>? where = null,
-		IEnumerable<Expression<Func<T, object>>>? orderBy = null, int? offset = null, int? count = null,
-		CancellationToken cancellationToken = default) =>
-		_repository.GetAsync(where, orderBy, offset, count, cancellationToken);
+    public virtual Task<T> GetAsync(Expression<Func<T, bool>>? where = null,
+        IEnumerable<Expression<Func<T, object>>>? orderBy = null, int? offset = null, int? count = null,
+        CancellationToken cancellationToken = default) =>
+        _repository.GetAsync(where, orderBy, offset, count, cancellationToken);
 
-	public virtual Task<IReadOnlyCollection<T>> GetManyAsync(Expression<Func<T, bool>>? where = null,
-		IEnumerable<Expression<Func<T, object>>>? orderBy = null, int? offset = null, int? count = null,
-		CancellationToken cancellationToken = default) =>
-		_repository.GetManyAsync(where, orderBy, offset, count, cancellationToken);
+    public virtual Task<IReadOnlyCollection<T>> GetManyAsync(Expression<Func<T, bool>>? where = null,
+        IEnumerable<Expression<Func<T, object>>>? orderBy = null, int? offset = null, int? count = null,
+        CancellationToken cancellationToken = default) =>
+        _repository.GetManyAsync(where, orderBy, offset, count, cancellationToken);
 
-	public virtual Task AddAsync(T entity, CancellationToken cancellationToken = default) =>
-		_repository.AddAsync(entity, cancellationToken);
+    public virtual Task AddAsync(T entity, CancellationToken cancellationToken = default) =>
+        _repository.AddAsync(entity, cancellationToken);
 
-	public virtual Task AddManyAsync(IReadOnlyCollection<T> entities,
-		CancellationToken cancellationToken = default) =>
-		_repository.AddManyAsync(entities, cancellationToken);
+    public virtual Task AddManyAsync(IReadOnlyCollection<T> entities,
+        CancellationToken cancellationToken = default) =>
+        _repository.AddManyAsync(entities, cancellationToken);
 
-	public virtual Task UpdateAsync(T item, CancellationToken cancellationToken = default) =>
-		_repository.UpdateAsync(item, cancellationToken);
+    public virtual Task UpdateAsync(T item, CancellationToken cancellationToken = default) =>
+        _repository.UpdateAsync(item, cancellationToken);
 
-	public virtual Task UpdateManyAsync(IReadOnlyCollection<T> entities,
-		CancellationToken cancellationToken = default) =>
-		_repository.UpdateManyAsync(entities, cancellationToken);
+    public virtual Task UpdateManyAsync(IReadOnlyCollection<T> entities,
+        CancellationToken cancellationToken = default) =>
+        _repository.UpdateManyAsync(entities, cancellationToken);
 
-	public virtual Task DeleteAsync(T entity, CancellationToken cancellationToken = default) =>
-		_repository.DeleteAsync(entity, cancellationToken);
+    public virtual Task DeleteAsync(T entity, CancellationToken cancellationToken = default) =>
+        _repository.DeleteAsync(entity, cancellationToken);
 
-	public virtual async Task DeleteManyAsync(IReadOnlyCollection<T> entities,
-		CancellationToken cancellationToken = default) =>
-		await _repository.DeleteManyAsync(entities, cancellationToken);
+    public virtual async Task DeleteManyAsync(IReadOnlyCollection<T> entities,
+        CancellationToken cancellationToken = default) =>
+        await _repository.DeleteManyAsync(entities, cancellationToken);
 }
